+++ tests/WalletTest.php
@@ -7,17 +7,43 @@ use Juspay\Model\Wallet;
 
 class WalletTest extends \PHPUnit_Framework_TestCase {
     private $customerTest;
+    private $wallet;
+    
+    public function testCreate() {
+        if ($this->customerTest == null){
+            $this->customerTest = new CustomerTest ();
+            $this->customerTest->testCreate ();
+        }
+        $wallet = Wallet::create ( $this->customerTest->customer->objectReferenceId, "MOBIKWIK" );
+        assert ( $wallet != null );
+    }
     public function testList() {
-        $this->customerTest = new CustomerTest ();
-        $this->customerTest->testCreate ();
+        $this->testCreate();
         $wallets = Wallet::listAll ( $this->customerTest->customer->objectReferenceId );
         assert ( $wallets != null );
     }
     public function testRefresh() {
-        $this->customerTest = new CustomerTest ();
-        $this->customerTest->testCreate ();
+        $this->testCreate();
         $wallets = Wallet::refresh ( $this->customerTest->customer->objectReferenceId );
         assert ( $wallets != null );
+    }
+    public function testRefreshByWalletId() {
+        $this->testCreate();
+        $wallet = Wallet::refreshByWalletId ( $this->wallet->id );
+        assert ( $wallet != null );
+    }
+    public function testCreateAndAuthenticate() {
+        if ($this->customerTest == null){
+            $this->customerTest = new CustomerTest ();
+            $this->customerTest->testCreate ();
+        }
+        $wallet = Wallet::createAndAuthenticate ( $this->customerTest->customer->objectReferenceId, "MOBIKWIK" );
+        assert ( $wallet != null );
+    }
+    public function testAuthenticate() {
+        $this->testCreate();
+        $wallet = Wallet::authenticate ( $this->wallet->id );
+        assert ( $wallet != null );
     }
 }
 require_once __DIR__ . '/TestEnvironment.php';